aquaponics_v3/raspberry_pi_camera/requirements.txt
aiohttp==3.9.1
python-dotenv==1.0.0
picamera2==0.3.17
opencv-python==4.8.0.76
libcamera==0.0.5
